separate stdlib into each module folder
