+++ src/docs/entries/stdlib/array/mod.rs
