+++ src/docs/entries/stdlib/math/mod.rs
