+++ src/docs/entries/stdlib/mod.rs
@@ -1,9 +1,9 @@
 //! Standard library documentation entries, one module per file.
-pub mod arrays;
-pub mod constants;
+pub mod array;
 pub mod fs;
 pub mod io;
 pub mod math;
+pub mod math_consts;
 pub mod path;
 pub mod process;
 pub mod random;
