+++ src/docs/entries/stdlib/random/mod.rs
