+++ src/docs/entries/stdlib/result/mod.rs
